cb_pcidas: Set len_chanlist for AO subdevice
Boards that support asynchronous commands on the AO subdevice (those
that have an AO FIFO) have code to support output to both channels on
each scan trigger (if both channels are in the channel list).  However,
the AO subdevice initialization code in cb_pcidas_attach() neglected to
set the subdevice's 'len_chanlist' member, leaving it set to 0.  The
core Comedi code changes it to 1, but that limits the command's channel
list to a single channel.  Change the initialization code to set it to
the number of AO channels for boards that support AO commands.

Signed-off-by: Ian Abbott <abbotti@mev.co.uk>