Do not build the serial2002 module if kernel cannot support it
The serial2002 module tries to access TTYs like a userspace program by
using the 'set_fs(KERNEL_DS)' trick.  That could probably be avoided for
read and write operations, but not for TTY ioctl calls that use a
pointer for the ioctl optional argumenti.

Support for 'set_fs()' is gradually being removed.  The the 5.10 kernel
it is being removed for at least the powerpc, riscv and x86
architectures, resulting in build errors for the serial2002 driver.

Add configuration checks for 'set_fs()' support and do not build the
serial2002 driver if there is no support.  The tests should work for
kernel versions up to 5.10, but may break for future kernel versions if
the CONFIG_SET_FS kernel configuration option introduced in the 5.10
kernel is removed altogether in a later kernel version.

Also add some code to 'dkms.conf' to avoid installing the module if it
cannot be built, and update the date in the PACKAGE_VERSION macro
therein.  This code may also break for kernel versions beyond 5.10 for
similar reasons to the above.

Signed-off-by: Ian Abbott <abbotti@mev.co.uk>