comedi/Makefile.am: Don't ignore errors from kernel's make modules
There is an old workaround in comedi/Makefile to invoke the kernel's
build system with errors ignored, which causes comedi's make to succeed.
Apparently, this was needed for NFS filesystems with root_squash.  I
don't know if that is still the case for modern systems (the workaround
was added by Frank Mori Hess in 2007), but it probably doesn't affect
too many people.  The confusion caused by comedi make appearing to
succeed even when it fails to build the modules probably affects more
people.

I have changed it not to ignore errors from the kernel's build system
and left a cryptic comment about it in comedi/Makefile.am.

Signed-off-by: Ian Abbott <abbotti@mev.co.uk>