drivers: Use 'fallthrough;' before case labels that fall through
For those drivers that deliberately omit a 'break;' statement in a
'switch' statement before a 'case' or 'default' label, add a
'fallthrough;' statement.  This was defined in kernel 5.4 and we have a
dummy kernel compatibility macro for it for earlier kernels.

Keep the existing "fall-through" or "FALLTHROUGH" type comments in the
code for use when building with earlier kernels that are using the dummy
'fallthrough' macro.

Signed-off-by: Ian Abbott <abbotti@mev.co.uk>