comedi_compat32: Move comedi_compat_ioctl() into comedi_fops.c
Move the bulk of the 32-bit compatibility code from comedi_compat32.c to
comedi_fops.c, leaving behind the ioctl mapping code for kernel 2.6.10
and earlier.

The 32-bit compatible ioctl code needs to be reworked for compatibility
with kernel version 5.15 and later, and comedi_fops.c is the best place
to make the changes.

Signed-off-by: Ian Abbott <abbotti@mev.co.uk>