comedi_fops: Make comedi_ioctl() a wrapper around comedi_unlocked_ioctl()
If HAVE_UNLOCKED_IOCTL is defined we use the unlocked_ioctl file handler
comedi_unlocked_ioctl(), otherwise we use the ioctl file handler
comedi_ioctl().  Only one of those functions was defined.  Change it so
that the comedi_unlocked_ioctl() function is always defined and change
comedi_ioctl() (if defined) to be a simple wrapper that calls
comedi_unlocked_ioctl().

Signed-off-by: Ian Abbott <abbotti@mev.co.uk>