comedi_fops: preserve long width return value for comedi_compat_ioctl()
comedi_compat_ioctl() eventually calls comedi_unlocked_ioctl() (unless
there is an error before the call).  comedi_unlocked_ioctl() returns a
'long' instead of an 'int', as does comedi_compat_ioctl(), so preserve
the 'long' value instead of converting it to an 'int' and back to a
'long'.  This doesn't really matter since the return value will fit in
an 'int' anyway, but is cleaner.

Signed-off-by: Ian Abbott <abbotti@mev.co.uk>