fixup! comedi_fops: Do not use compat_alloc_user_space() (fix for kernel 5.15)
It seems I did not commit all of the changes required to eliminate
compat_alloc_user_space(), leaving the code in an unbuildable state.
Specifically, I the code for do_rangeinfo_i() was missing.  Here is the
remaining part of the change.

Thanks to GitHub user 'warr' for reporting the error.

Signed-off-by: Ian Abbott <abbotti@mev.co.uk>