comedi_fops: Do not fail buffer resize if previous resize failed
The resize_async_buffer() function returns an error -EINVAL if
s->prealloc_buf is NULL.  That could have resulted from a previous
failure to allocate the buffer.

Allow the old s->prealloc_buf to be NULL so that the user can retry the
buffer configuration (perhaps with a smaller buffer size) if it fails.

Signed-off-by: Ian Abbott <abbotti@mev.co.uk>