m4/as-linux.m4: Fix COMEDI_CHECK_HAVE_SET_FS() for 5.18 kernel
The 5.18 kernel removes the CONFIG_SET_FS option introduced in the 5.10
kernel and removes support for 'set_fs()' altogether.  Removal of the
CONFIG_SET_FS option in the 5.18 kernel broke the tests performed by
COMEDI_CHECK_HAVE_SET_FS(), making it think the kernel supported
'set_fs()'.  Update the test to also check for the presence of the
'<asm-generic/access_ok.h>' header that was added in 5.18.

Signed-off-by: Ian Abbott <abbotti@mev.co.uk>