inc-wrap/asm/io.h: Try and make isa_virt_to_bus() compile
Only a few architectures define isa_virt_to_bus() as a macro or an
inline function.  For those that don't, replace isa_virt_to_bus with
virt_to_phys (not virt_to_bus, because that is going away in Linux
kernel 6.0).

We cannot check if isa_virt_to_bus is defined as an inline function, so
use some arch-specific and kernel version-specific educated guesswork.

Really, we should modify drivers to use the ISA DMA API if the kernel
provides it, but the isa_virt_to_bus stuff would still be needed for
older kernels.

Signed-off-by: Ian Abbott <abbotti@mev.co.uk>