adl_pci9118: Avoid a -Wunused-but-set-parameter compiler warning
In setup_channel_list(), the `rot` parameter may get assigned to, but
the value only gets used afterwards if the `PCI9118_PARANOIDCHECK` macro
is defined.  Add a dummy `(void)rot;` statement to avoid the warning.
It's simpler than wrapping it in an `#ifdef PCI9118_PARANOIDCHECK` /
`#endif` pair and should get optimized out anyway.

Signed-off-by: Ian Abbott <abbotti@mev.co.uk>