inc-wrap/linux/mutex.h: Avoid incorrectly defining dummy mutex_destroy()
mutex_destroy() has not been defined as a macro since Linux version 3.1,
so the test for a Redhat backported <linux/mutex.h> without
mutex_destroy() is broken.  In the mainline kernel, mutex_destroy() has
been defined as a macro and/or function since kernel version 2.6.16, so
do not define it for kernel version 2.6.16 or later.

Signed-off-by: Ian Abbott <abbotti@mev.co.uk>