comedi_fops: Use dma_mmap_coherent() to map DMA buffer pages
The comedi buffer is allocated as individual pages.  If allocating for
DMA use, the pages are allocated individually using
dma_alloc_coherent().

In comedi_mmap(), the buffer pages are mapped individually using
remap_pfn_range().  However, for pages allocated using
dma_alloc_coherent(), the pages ought to be mapped using
dma_mmap_coherent() to avoid abusing the DMA API (in particular the
calling of virt_to_page() on the blocks allocated by
dma_alloc_coherent()).

The problem is that dma_mmap_coherent() maps the whole VMA to a
contiguous range of DMA coherent memory pages, but our buffer is split
into individual pages.  To work around this, use a neat trick found on
Stackoverflow[1] that involves modifying the VMA vm_start and vm_end
addresses before calling dma_mmap_coherent() for each page, and
restoring the old vm_start and vm_end addresses afterwards.

[1] https://stackoverflow.com/a/67220955/5264491

Outstanding issue: The allocation of the pages for the DMA buffer still
uses virt_to_page() to construct a list of pages prior to calling vmap()
to map a contiguous range of virtual address space to the buffer pages.

Signed-off-by: Ian Abbott <abbotti@mev.co.uk>