comedi_fops: COMEDI_BUFINFO: return error if no active command
The COMEDI_BUFINFO ioctl is used to advance the current position in the
buffer and/or het the current buffer position.  If no asynchronous
command is active for the file object that issued this ioctl, the
information is meaningless.  Change it to return an error -EINVAL in
this case.  Prior to this change, if a command was started via a
different file object, the ioctl returned -EACCES, but now it will
return -EINVAL, which is consistent with the current behavior of the
"read" and "write" file operation handlers.  Also change the logic for
returning errors and becoming non-busy to match that in the in-kernel
version of comedi.

Signed-off-by: Ian Abbott <abbotti@mev.co.uk>