ni_pcidio: Do not bother filling buffer with 0xaa byte values
On buffer change, the driver informs the mite module about the buffer
change and then it fills the buffer with byte value 0xaa using memset().
Do not bother filling the buffer.

Aiming to get remove the mapping of the buffer to vmalloc address range
in comedi, and just use the buffer pages directly, so there will no
longer be an overall mapping of the buffer to contiguous virtual address
space.  We could add a comedi_buf_memset() function to fill the buffer
with a byte value, but it's not worth it in this case.

Signed-off-by: Ian Abbott <abbotti@mev.co.uk>