dt282x: Change test for missing buffer in DMA interrupt routines
Replace the test for null s->async->prealloc_buf with a test for zero
s->async->prealloc_bufsz.

The prealloc_buf member in comedi points to the start of the comedi
buffer mapped to vmalloc address space.  Aiming to get rid of the
mapping to vmalloc address space and use individual buffer pages
directly.