kcomedilib: Remove the comedi_map() and comedi_unmap() functions
Breaking change in kcomedilib.

The comedi_map() function provided by the kcomedilib module gets a copy
of the prealloc_buf pointer that provides a linear mapping of the comedi
buffer.

I am aiming to get rid of the linear mapping of the buffer that is
accessed through the prealloc_buf pointer, and the comedi_map() function
is getting in the way.

Remove the comedi_map() and comedi_unmap() functions and hope no-one is
using them.

Signed-off-by: Ian Abbott <abbotti@mev.co.uk>