comedi: Remove the linear mapping of the comedi buffer (prealloc_buf)
Remove the mapping of the comedi buffer into a linear range of addresses
in vmalloc space pointed to by prealloc_buf.  This saves on using up
vmalloc address space, and also avoids abusing the DMA API that forbids
calling virt_to_page() on buffers allocated by dma_alloc_coherent().

The only unfortunate casualty of the whole exercise was the removal of
the comedi_map() and comedi_unmap() functions in the kcomedilib module.

Signed-off-by: Ian Abbott <abbotti@mev.co.uk>