inc-wrap/linux/timer.h: Fix for Linux kernel 6.16 compatibility
Linux kernel 6.16 renamed the `from_timer` macro to
`timer_container_of`, but we tested for the presence of `from_timer` to
decide whether to define the `timer_setup` function for compatibility
with kernels prior to 4.14.

Only define the `timer_setup` function if neither `timer_container_of`
nor `from_timer` was defined.

Define `timer_container_of` if missing.

Keep the definition of the `from_timer` macro if missing because some
Comedi drivers are still using it (but not for much longer!).

Signed-off-by: Ian Abbott <abbotti@mev.co.uk>