inc-wrap/linux/timer.h: Remove compatibility definition of from_timer()
All the Comedi drivers that called `from_timer()` have been updated to
call `timer_container_of()` instead.  We defined a `from_timer()`
macro for compatibility if it was not already defined, but that is no
longer required.  We still test for the existence of `from_timer()` as
part of the test for defining the `timer_setup()` compatibility
function.

Signed-off-by: Ian Abbott <abbotti@mev.co.uk>