Fix for Red Hat back-port of class_create() from 6.4 to 5.19
Red Hat back-ported the version of class_create() from mainline kernel
version 6.4 to their later 5.14 kernels for RHEL9, which breaks comedi's
wrapper code for this function that redefines class_create() as a macro
to resemble the 6.4 version.

The back-port can be detected by the fact that class_create is not
defined as a macro.  Change the compatibility wrapper code to detect
this and avoid wrapping the function if it has been back-ported from
6.4.

Signed-off-by: Ian Abbott <abbotti@mev.co.uk>