m4/as-linux.m4: Remove the --with-machine=MACHINE option
Nothing really uses the --with-machine=MACHINE option.  It was intended
to be a `uname -m` value used to check .config files, but that was never
implemented.

Also remove the configuration code that tries to map an RPM target
architecture to a MACHINE (uname -m) value, as that could make the
configuration script fail with an error if it did not recognize the RPM
target architecture as belonging to a limited set of x86 architectures.