comedi/Makefile.am: Remove uninstall option
Using `find` to find and remove the kernel modules could be dangerous if
the comedi module directory has a generic name like "extra" or
"updates".  Also, the module files might be compressed so that the search
for '*.ko' files might fail in that case.  Also, 'depmod' ought to be run
on the top-level modules directory after removing modules.  For those
reasons, decline to uninstall the module files.

The original reason for removing the files was for passing 'make
distcheck' tests, but we can work around that later.

Signed-off-by: Ian Abbott <abbotti@mev.co.uk>