Workaround a problem with make distclean running in parallel
When `make distclean` is running with MAKEFLAGS=j, and is cleaning in
parallel, the comedi/comedi_kbuild.inc was being deleted in parallel
with the running of the kernel's make clean to delete the modules, so
that the file was missing when one of the Kbuild files tried to
`include` it.

Fix it by renaming comedi/comedi_build.inc.in to
comedi/comedi_kbuild_copy.inc.in, and adding a rule in
comedi/Makefile.in to create comedi_kbuild.inc from
comedi_kbuild_copy.inc by copying it. (comedi_kbuild_copy.inc is created
by the ./configure script from comedi_kbuild_copy.inc.in.)

Signed-off-by: Ian Abbott <abbotti@mev.co.uk>