configure.ac, comedi/Makefile.am: Add `--enable-modulesdir-prefix` option
Normally, the configured prefix (like "/usr/local" or "/usr" should be
ignored for Linux kernel modules.  The configure script does set the
prefix to "/" if it has not been configured explicitly.  Rather than do
that, add an option `--enable-modulesdir-prefix` to enable the module
installation path to be prefixed.  When disabled (the default), there
will be no prefix.  (However $(DESTDIR) is still supported for staged
installation.)

One circumstance where kernel modules need to be installed with the
prefix is when doing `make distcheck`.  To allow that, configure.ac sets
`DISTCHECK_CONFIGURE_FLAGS="--with-modulesdir=\\\$\$\(prefix\)/modules"`.
Change that to `DISTCHECK_CONFIGURE_FLAGS="--enable-moduledir-prefix"`.

This change will help `rpmbuild` when it is supported, because it
configures the prefix but does not expect kernel modules to be installed
to the prefix.

Signed-off-by: Ian Abbott <abbotti@mev.co.uk>