m4/as-linux.m4: Remove `--with-modulesdir` and `--with-modulesdeveldir` options
Nothing is installed to the modulesdeveldir, so remove the
`--with-modulesdeveldir` option.

Modules are installed via the kernel Makefiles, and the only influence
over the installation location is via the variables `INSTALL_MOD_PATH`
(which sets an installation prefix for staged installation), and
`INSTALL_MOD_DIR` (which sets a postfix directory name such as
"/extra").  The `--with-modulesdir` option is trying to set the full
installation path within `$(DESTDIR)`, including the `$(uname -r)`
value, which cannot be done.  So remove the `--with-modulesdir` option.

Signed-off-by: Ian Abbott <abbotti@mev.co.uk>