INSTALL.md, README.md: Update for modern systems
The instructions in INSTALL.md are pretty obsolete for use on modern
Linux systems.  Preserve it as INSTALL-old.md and rewrite it for modern
systems.  Also update the README.md file.

The new INSTALL.md has an "Automatic Installation" section that has yet
to be written.  It will have instructions for using Akmods packages on
Red Hat systems, or DKMS on other systems.

Signed-off-by: Ian Abbott <abbotti@mev.co.uk>