dkms.conf.in: Use `--disable-maintainer-mode` configure script option
The `dkms add` command does not preserve timestamps when copying a
source directory (possibly from an unpacked tarball) (at least for DKMS
version 3.2.0).  This can result in build failures if Automake's
"maintainer-mode" is enabled, because it may try and update the
Makefiles, which relies on Automake being installed.

Currently, we enable "maintainer-mode" by default, so explicitly disable
it when running the configure script from DKMS.

Signed-off-by: Ian Abbott <abbotti@mev.co.uk>