8255: Fail standalone 8255 attach if any I/O region unavailable
Configuring a standalone 8255 device with one or more DIO subdevices,
each specified by its own I/O port address current succeeds if any of
the I/O port address regions are unavailable, with the subdevice being
configured as an "unused" subdevice in that case.  That behavior is a
bit confusing as it looks like the device is configured OK even though
one or more subdevices may be unusable.  Change the behavior to fail the
attachment if any of the I/O port regions are unavailable.

Signed-off-by: Ian Abbott <abbotti@mev.co.uk>