kcomedilib: Add loop checking open and close functions
Add comedi_open_from(), comedi_open_minor_from(), and
comedi_close_from() functions.  These are like the comedi_open(),
comedi_open_minor(), and comedi_close() functions but have an extra
`from` parameter indicating the minor device number of the COMEDI device
making the request if the `from` parameter value is in the range of
COMEDI board minor device numbers.  Otherwise, the `from` parameter is
ignored.

If the `from` parameter is in range, the `comedi_open_from()` and
`comedi_open_minor_from()` functions will return NULL if a cycle of
COMEDI devices that mutually open each other would be created.  If
`comedi_open_from()` or `comedi_open_minor_from()` is called with a
`from` parameter that is in range, and the COMEDI device is successfully
opened, then `comedi_close_from()` should be called with the same `from`
parameter value to close the device, in order to maintain the data used
to check for cycles.

Signed-off-by: Ian Abbott <abbotti@mev.co.uk>