kcomedilib: Don't get and put modules
When comedi_close() or comedi_close_from() is called, the device being
closed might be detached or might not be attached to the same driver as
it was when opened.  In that case the call to module_put() when closing
would lead to undefined behavior.  Remove all the calls to
`try_module_get()` and `module_put()` from kcomedilib.  (The calls would
make sense if COMEDI devices were prevented from being detached while
they have been opened by `comedi_open()`, etc.)

Signed-off-by: Ian Abbott <abbotti@mev.co.uk>