inc-wrap/kernel.h: Define KERN_CONT and KERN_DEFAULT if undefined
For use with printk(), the KERN_CONT macro was first defined in Linux
kernel version 2.6.24, and the KERN_DEFAULT macro was first defined in
Linux kernel 2.6.31.  (Later kernels moved them to the linux/printk.h
and later still to the linux/kern_levels.h header, but they still get
pulled in by `#include <linux/kernel.h>`.)  Define them to expand to
empty string literals if they are not defined.

Signed-off-by: Ian Abbott <abbotti@mev.co.uk>