m4/as-linux.m4: Look for Makefile variable anywhere in kernel source path
In the check for separate Linux source and build directories in
AS_LINUX_SRC_DIR() that checks the Makefile in the build directory for
references a source directory, if it ends up with a potential source
directory string beginning with `$(` it is disregarded.  Change that so
that it disregards a potential source directory containing `$(` anywhere
in the string, not just at the beginning.