inc-wrap: Remove linux/version.h wrapper
There was a mechanism added in 2002 to allow UTS_RELEASE to be redefined
to a supplied UTS_RELEASE_OVERRIDE value.  I don't know why or how it
was intended to be used, but it's pretty old (pre 2.6 kernel) and
probably irrelevant now, so remove it.  Since that is the only thing in
the linux/version.h wrapper, remove the whole wrapper.

Signed-off-by: Ian Abbott <abbotti@mev.co.uk>