das16: Use generic DMA API instead of PCI DMA API
Since Linux kernel versions prior to 2.6 are no longer supported, use
the generic DMA API functions instead of the PCI DMA API wrappers.

NOTE: This driver's use of the PCI DMA API was horribly broken due to
invalid use of a null PCI device pointer, and its use of the DMA API is
still horibly broken due to use of a null device pointer, so it has been
disabled.  Device attachment will fail with an error if configured to
use a DMA channel.  This may be fixed in the future.

Signed-off-by: Ian Abbott <abbotti@mev.co.uk>