inc-wrap/linux/delay.h: Avoid conflicts with backported kernel functions
Debian's 2.6.8 kernel backports the ssleep() function from the mainline
2.6.9 kernel, resulting in a compilation error when the function is
redefined by our kernel compatibility wrapper.  Rename the function to
comedi_ssleep() in our wrapper and define ssleep(seconds) as a macro
that calls it.  Also rename msleep() to comedi_msleep() and define
msleep(msecs) as a macro that calls it.  Also rename
msleep_interruptible() to comedi_msleep_interruptible() and define
msleep_interruptible(msecs) as a macro that calls it.

Signed-off-by: Ian Abbott <abbotti@mev.co.uk>