cb_pcidac: Replace ioread/write8/16 calls
The ioread8(), iowrite8(), ioread16(), and iowrite16() functions were
were not introduced until kernel version 2.6.9, so replace calls to them
with calls to readb(), writeb(), readw(), and writew() for compatibility
with earlier kernels.

NOTE: It is possible that the DIO register access in PCI BAR 3 is
broken.  I don't know whether the author tested the DIO subdevices, but
according to the "Register Map for the PCI-DAC6700 Series" document,
RegMapPCI-DAC670x.pdf, PCI BAR 3 is documented as being I/O mapped, but
the driver currently assumes they are memory mapped.

Signed-off-by: Ian Abbott <abbotti@mev.co.uk>