inc-wrap/linux/module.h: Define COMEDI_MODNAME
In kernel versions since 2.6.16, the KBUILD_MODNAME macro expands to a
string literal, but before 2.6.16 it expanded to an identifier.

Define COMEDI_MODNAME to expand to a string literal, expanding to
KBUILD_MODNAME for kernels since 2.6.16, or to a stringified version of
KBUILD_MODNAME for kernels prior to 2.6.16.

Modules can use COMEDI_MODNAME instead of KBUILD_MODNAME when they need
the module name as a string literal.

Signed-off-by: Ian Abbott <abbotti@mev.co.uk>