inc-wrap/linux/printk.h: New compatibility header
New compatibility header.  It includes the real <linux/printk.h> for
kernel version 2.6.37 or later, which is when it first appeared.  It is
also included by our linux/kernel.h header, which is where
printk-related stuff was defined for kernels prior to 2.6.37.

It defines or redefines macros such as pr_err(), pr_warning(), etc.,
that act as wrappers around the printk() function.  Versions of these
macros from kernel versions 2.6.24 to 2.6.27.x inclusive are redefined
to use the pr_fmt(fmt) macro that can be defined by driver code, but
defaults to just expanding to its parameter.

The compatibility definitions of KERN_CONT and KERN_DEFAULT have been
moved from the linux/kernel.h compatibility header to this one.

Signed-off-by: Ian Abbott <abbotti@mev.co.uk>