inc-wrap/linux/usb.h: Redefine usb_unlink_urb() to complete asynchronously
For kernel versions prior to 2.6.8, redefine usb_unlink_urb() so that it
sets the URB's URB_ASYNC_UNLINK transfer flag before calling the real
usb_unlink_urb().  This will cause it to return before the URB has been
completed, which matches the modern usage.

Code should call usb_kill_urb() instead to unlink and wait for the
transfer to complete.

Signed-off-by: Ian Abbott <abbotti@mev.co.uk>