Change parameters of comedi_usb_auto_config() and comedi_usb_auto_unconfig()
Change the type of the first parameter of both functions from `struct
usb_device *` to `struct usb_interface *`.

Replace the `const char *board_name` parameter of
`comedi_usb_auto_config()` with a `comedi_driver *driver` parameter and
an `unsigned long context` parameter.  The `context` value is currently
unused by the function.  The "usbdux", "usbduxfast", and "usbsuxsigma"
drivers set the context to 0.  The "vmk80xx" driver sets the context to
the model number, which came from the `driver_info` member of the
`struct usb_device_id` when the driver's USB interface "probe" function
was called.

Signed-off-by: Ian Abbott <abbotti@mev.co.uk>