comedi: Change comedi_alloc_board_minor() to return a comedi_device* and lock its mutex
Change comedi_alloc_board_minor() to return a pointer to a comedi_device
with its mutex pre-locked on success, and to return an ERR_PTR() value
on failure.  Previously, it returned the minor device number of the
allocated device without locking the device's mutex.

Callers need to unlock the mutex on success, but do not need to lock it
separately, so the auto-config code does not need to check if the device
is attached after locking the mutex.

Signed-off-by: Ian Abbott <abbotti@mev.co.uk>