drivers.c: Remove 'attached' check in comedi_device_attach_driver()
comedi_device_attach_driver() is called from comedi_device_attach(),
which has already checked the comedi_device is not attached, and from
comedi_old_auto_config(), which dynamically allocates a comedi_device
that it not attached.  Remove the redundant code from
comedi_device_attach_driver() that returns an error if the comedi_device
is attached.

Signed-off-by: Ian Abbott <abbotti@mev.co.uk>