drivers.c: Refactor comedi_device_attach_driver()
Factor new function `comedi_device_attach_driver_wrapper()` out of
`comedi_device_attach_driver()`.  It has a `void *context` pointer
pointing to a `struct comedi_device_attach_wrapper_context` that has
members `comedi_devconfig *it` and `bool *matched` that are copies of
the `comedi_devconfig *it` and `bool *matched` parameters from
`comedi_device_attach_driver()`.  It checks that the driver has a match
for `it->board_name`, returning `-EINVAL` if it doesn't match, or
setting `*matched` to `true` if it does match.  It also checks that the
driver has an `attach` handler, returning `-EINVAL` if it doesn't.  It
then sets the device's `driver` member to point to the `comedi_driver`
and calls the driver's `attach` returning the result.  The original
`comedi_device_attach_driver()` function is responsible for incrementing
the driver module's reference count, calling the new wrapper function,
doing the post-config cleanup, and decrementing the driver module's
reference count on failure.  (It remains incremented on success.)

The new `comedi_driver_attach_driver_wrapper()` function is also
intended for use by comedi_old_auto_config() in a later patch.

Signed-off-by: Ian Abbott <abbotti@mev.co.uk>