Add new auto-config mechanism
Allow drivers to define an `auto_attach()` handler.  It differs from the
`attach()` handler in that the comedi device's `hw_dev` member is
already set to the hardware device (from which the enclosing PCI device
or USB interface can be determined by the handler), the `board_ptr`
member is always NULL, but can be set by the handler, there is no
options list, but there is unsigned long `context` value that came from
the driver's call to `comedi_pci_auto_config()`,
`comedi_usb_auto_config()`, or the new, generic `comedi_auto_config()`
function.  The previously static `comedi_auto_unconfig()` function in
drivers.c is now exported for use by modules that called
`comedi_auto_config()`.

The `comedi_pci_auto_config()` and `comedi_usb_auto_config()` functions
will call the driver's `auto_attach()` handler if it has one, and fall
back to using the `attach()` handler in the same way that it used to.

Signed-off-by: Ian Abbott <abbotti@mev.co.uk>