comedi_fops.c: Add comedi_clear_board_dev()
Add local function `comedi_clear_board_dev()` as an alternative to
`comedi_clear_board_minor()` when we already have a pointer to the
`comedi_device`.  It uses the board minor device number stored in the
`comedi_device` (which must have already been initialized) and only
clears the entry in `comedi_board_minor_table[]` if it points to the
specified `comedi_device`.  Rather than returning the old table entry,
it returns `true` if the entry matched (and so has just been cleared),
and returns `false` otherwise.

Call `comedi_clear_board_dev()` instead of `comedi_clear_board_minor()`
in `comedi_unlocked_ioctl()` in the code that frees a dynamically
allocated comedi device detached by the `COMEDI_DEVCONFIG` ioctl.  That
ought to return `true` but check it just in case before freeing the
device.

Signed-off-by: Ian Abbott <abbotti@mev.co.uk>