comedi_fops.c: Remove check for CONFIG_KMOD for auto-loading modules
The `CONFIG_KMOD` option changed to `CONFIG_MODULES` in kernel version
2.6.27.  If it is not configured, the `request_module()` call will just
return an error, which is OK because we ignore it.  Therefore, just
ignore the config option test and call `request_module()` regardless
when attempting to request the module for "char-major-98-N" (where "N"
is the minor device number and "98" is the major device number).

Signed-off-by: Ian Abbott <abbotti@mev.co.uk>