Use a rw_semaphore to protect against device detachment
The 'read' and 'write' file operations on comedi devices do not use the
main mutex in the `comedi_device` to avoid contention with ioctls that
may take a while to complete.  However, it is necessary to protect
against the device being detached while the operation is in progress.
Add member `struct rw_semaphore attach_lock` to `comedi_device` for this
purpose and initialize in on creation.  Use it in the code for attaching
and detaching devices and in the 'read' and 'write' handlers.  Also use
it for the 'poll' handler instead of the mutex.

Need to be careful to avoid deadlocks in some places.  The main mutex
cannot be locked while `attach_lock` is locked.  Also need to avoid
waiting to lock `attach_lock` in the 'mmap' operation, so use
`try_down_read()` instead of `down_read()` in that case.

Signed-off-by: Ian Abbott <abbotti@mev.co.uk>