comedi_fops.c: Some changes to poll operation
When handling the "poll" file operation, check more conditions for
marking the device readable or writeable when reading or writing would
result in an error.  If an asynchronous command running on the 'read'
subdevice was set up by a different file, or the command was set up in
the 'write' direction, then mark the polling as readable as reading
would return an error.  Similarly, if an asynchronous command running on
the 'write' subdevice was set up by a different file, or the command was
set up in the 'read' direction, then mark the polling file as writeable
as writing would return an error.  Also, do not allocate space from the
buffer for writing, just check that space is available for writing.
Allocating the buffer space before checking a command was set up in the
'write' direction can cause problems if the subdevice supports commands
in either direction and currently has a running command in the 'read'
direction.

Signed-off-by: Ian Abbott <abbotti@mev.co.uk>