comedi_fops.c: Rearrange comedi_write() code
Move the code that makes the subdevice become non-busy to a point after
the `while` loop, using a new `become_nonbusy` flag variable set inside
the `while` loop to decide whether it needs to be done.  There is now a
single place where the task is removed from the wait queue, so remove
the `on_wait_queue` flag variable.

Signed-off-by: Ian Abbott <abbotti@mev.co.uk>