comedidev.h: Add some sample size manipulation functions
Add some static inline helper functions:

`comedi_bytes_per_sample(subd)` is the same as the existing
`bytes_per_sample(subd) and determines the size of a comedi sample in
bytes.

`comedi_sample_shift(subd)` determines the log2 of the comedi sample
size, so it can be used in bit-shift operations to multiply or divide by
the sample size.

`comedi_bytes_to_samples(subd, nbytes)` converts a number of bytes to a
number of samples, rounding down.

`comedi_samples_to_bytes(subd, nsamples)` converts a number of samples
to a number of bytes.

Signed-off-by: Ian Abbott <abbotti@mev.co.uk>