comedi_fops.c: Correct CMDF_WRITE in user command if necessary
For subdevices that support support commands in only the "read"
direction, or only the "write" direction, force the CMDF_WRITE flag to
the correct value in the copy of the comedi_cmd copied back to the user,
and the copy of the comedi_cmd in the comedi_async structure.  But do
not treat incorrect setting of the CMDF_WRITE flag as an error.

This also means that the CMDF_WRITE flag in the comedi_async structure's
copy of comedi_cmd can be tested to determine the direction of the
command.  the comedi_poll() function already does that.

Signed-off-by: Ian Abbott <abbotti@mev.co.uk>