comedi_fops.c: Use command direction for kill_fasync()
In comedi_event(), use the state of the CMDF_WRITE flag in the
comedi_command stored in the comedi_async structure to determine whether
to use POLL_OUT or POLL_IN for the last argument of the call to
kill_fasync().  This relies on the the CMDF_WRITE flag being set
correctly, but the the COMEDI_CMD and COMEDI_CMDTEST ioctls were changed
recently to ensure that.  Previously, comedi_event() would call
kill_fasync() twice for subdevices that supported commands in both the
"read" and "write" direction, but we should only call it once, depending
on the actual direction of the command.

Signed-off-by: Ian Abbott <abbotti@mev.co.uk>