Use reference counted buffers
Use a reference counted buffer allocation controlled by a struct
comedi_buf_map.  It contains a `struct kref refcount` member used to
only free it when its reference count goes to 0.  There are
`comedi_buf_map_get()` and `comedi_buf_map_put()` functions to
manipulate the reference count.  The comedi_async structure now has a
`bm` member pointing to one of these, and this pointer only changes
while the subdevice's spin-lock is locked.  The change to the buffer
storage required a minor change to the "mite" driver when it is filling
in its DMA ring buffer descriptors.

The functions in "comedi_fops.c" that copy memory between the buffer and
user-space now allow wraparound, which simplifies the wraparound
handling in `comedi_read()` and `comedi_write()`.

Also implement some more changes to `comedi_read()` and `comedi_write()`
to return an error if the asynchronous command has not been set up, to
also allow zero-size reads and writes to return errors, and to check the
device is still in a suitable state to become non-busy after unlocking
`attach_lock` and locking `mutex`.

Also implement the `access` VM operation to allow mmapped buffers to be
accessed by debuggers such as GDB.  It is only implemented for kernel
version 2.6.27 or later.

These changes have been mostly borrowed from the version of comedi in
the Linux kernel source tree.

Signed-off-by: Ian Abbott <abbotti@mev.co.uk>