Another fix for "poll" operation in write direction
The "comedi_fops.c: Some changes to poll operation" change on 2025-12-04
removed a call to comedi_buf_write_alloc(), so it should now be checking
the value returned by comedi_buf_write_n_available() instead of the
value returned by comedi_buf_write_n_allocated().  But we also need to
fix the currently unused comedi_buf_write_n_allocated() function so that
it uses `buf_write_count` instead of `buf_write_alloc_count`.  Also
remove the line that rounds the number of available bytes down to a
multiple of the number of bytes per sample.  This all matches what has
been done in the version of Comedi in the Linux kernel sources.

Signed-off-by: Ian Abbott <abbotti@mev.co.uk>