comedi_buf.c: Add comedi_buf_write_samples() and comedi_buf_read_samples()
New exported function comedi_buf_write_samples() is similar to the
cfc_write_array_to_buffer() function in the "comedi_fc" module except
that the final parameter is a number of samples rather than a number of
bytes.  The return value is the number of bytes written to the buffer
(not the number of samples), and it checks for buffer overrun, advances
scan progress, and sets a pending COMEDI_CB_BLOCK event like
cfc_write_array_to_buffer().  A minor difference is that
cfc_write_array_to_buffer() writes nothing to the buffer if an overflow
would occur, whereas comedi_buf_writes_samples() will write as many
samples as will fit.

New exported function comedi_buf_read_samples() is similar to the
cfc_read_array_from_buffer() function in the "comedi_fc" module except
that the final parameter is the maximum number of samples to read rather
than the maximum number of bytes to read.  The return value is the
number of bytes read from the buffer (not the number of samples), and it
advances scan progress and sets a pending COMEDI_CB_BLOCK event like
cfc_read_array_from_buffer().

These functions were taken from the version of Comedi in the Linux
kernel sources.

Signed-off-by: Ian Abbott <abbotti@mev.co.uk>